Adding the Image classifier using convolution neural networks
In this commit, added the code to classify images from CIFAR-10 dataset. The dataset consists of airplanes, dogs, cats, and other objects. I have pre-processed the images and then add code to train a convolutional neural network on all the samples. The images are normalized and labels are one-hot encoded. The overrall neural network consists of convolutional, max pooling, dropout, and fully connected layers.